Refactor InjectedTextEdting and InjectedFormField
